# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vertiche LegalOps API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/tiendas/{id}/tramites": {
            "post": {
                "tags": [
                    "tiendas"
                ],
                "summary": "Create Tramite For Tienda",
                "operationId": "create_tramite_for_tienda",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TramiteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tramite"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tramites": {
            "get": {
                "tags": [
                    "tramites"
                ],
                "summary": "List Tramites",
                "operationId": "list_tramites",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Search"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Estado"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tipo"
                        }
                    },
                    {
                        "name": "estado_geografico",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Estado Geografico"
                        }
                    },
                    {
                        "name": "solo_vencidos",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Solo Vencidos"
                        }
                    },
                    {
                        "name": "por_vencer_dias",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Por Vencer Dias"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_Tramite_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tramites/{id}": {
            "get": {
                "tags": [
                    "tramites"
                ],
                "summary": "Get Tramite",
                "operationId": "get_tramite",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tramite"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "tramites"
                ],
                "summary": "Update Tramite",
                "operationId": "update_tramite",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TramiteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tramite"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AlertaSeveridad": {
                "type": "string",
                "enum": [
                    "info",
                    "warning",
                    "critical"
                ],
                "title": "AlertaSeveridad"
            },
...
            "CampoExtraidoDocumento": {
                "properties": {
                    "value": {
                        "type": "string",
                        "title": "Value"
                    },
                    "confidence": {
                        "type": "number",
                        "title": "Confidence"
                    }
                },
                "type": "object",
                "required": [
                    "value",
                    "confidence"
                ],
                "title": "CampoExtraidoDocumento"
            },
...
            "Documento": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "tramite_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tramite Ids"
                    },
                    "tramite_nombres": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tramite Nombres"
                    },
                    "nombre_archivo": {
                        "type": "string",
                        "title": "Nombre Archivo"
                    },
                    "ruta_archivo": {
                        "type": "string",
                        "title": "Ruta Archivo"
                    },
                    "estado_ocr": {
                        "$ref": "#/components/schemas/EstadoOCR"
                    },
                    "datos_extraidos": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/CampoExtraidoDocumento"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Datos Extraidos"
                    },
                    "requiere_revision_manual": {
                        "type": "boolean",
                        "title": "Requiere Revision Manual"
                    },
                    "cargado_por": {
                        "type": "string",
                        "title": "Cargado Por"
                    },
                    "cargado_por_nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cargado Por Nombre"
                    },
                    "cargado_en": {
                        "type": "string",
                        "title": "Cargado En"
                    },
                    "tienda_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tienda Id"
                    },
                    "tienda_nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tienda Nombre"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "tramite_ids",
                    "nombre_archivo",
                    "ruta_archivo",
                    "estado_ocr",
                    "requiere_revision_manual",
                    "cargado_por",
                    "cargado_en"
                ],
                "title": "Documento"
            },
...
            "EntidadTipo": {
                "type": "string",
                "enum": [
                    "alerta",
                    "documento",
                    "tienda",
                    "tramite",
                    "usuario"
                ],
                "title": "EntidadTipo"
            },
            "EstadoOCR": {
                "type": "string",
                "enum": [
                    "procesando",
                    "completado",
                    "baja_confianza",
                    "error"
                ],
                "title": "EstadoOCR"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HistorialItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "entidad_tipo": {
                        "$ref": "#/components/schemas/EntidadTipo"
                    },
                    "entidad_id": {
                        "type": "string",
                        "title": "Entidad Id"
                    },
                    "accion": {
                        "type": "string",
                        "title": "Accion"
                    },
                    "usuario_id": {
                        "type": "string",
                        "title": "Usuario Id"
                    },
                    "usuario_nombre": {
                        "type": "string",
                        "title": "Usuario Nombre"
                    },
                    "fecha": {
                        "type": "string",
                        "title": "Fecha"
                    },
                    "detalle": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Detalle"
                    },
                    "valor_anterior": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Valor Anterior"
                    },
                    "valor_nuevo": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Valor Nuevo"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "entidad_tipo",
                    "entidad_id",
                    "accion",
                    "usuario_id",
                    "usuario_nombre",
                    "fecha"
                ],
                "title": "HistorialItem"
            },
...
            "Observacion": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "descripcion": {
                        "type": "string",
                        "title": "Descripcion"
                    },
                    "severidad": {
                        "$ref": "#/components/schemas/AlertaSeveridad"
                    },
                    "fecha": {
                        "type": "string",
                        "title": "Fecha"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "descripcion",
                    "severidad",
                    "fecha"
                ],
                "title": "Observacion"
            },
...
            "PaginatedResponse_Tramite_": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/Tramite"
                        },
                        "type": "array",
                        "title": "Data"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "total_pages": {
                        "type": "integer",
                        "title": "Total Pages"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "total",
                    "page",
                    "page_size",
                    "total_pages"
                ],
                "title": "PaginatedResponse[Tramite]"
            },
...
            "Tramite": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "tienda_id": {
                        "type": "string",
                        "title": "Tienda Id"
                    },
                    "tienda_nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tienda Nombre"
                    },
                    "nombre": {
                        "type": "string",
                        "title": "Nombre"
                    },
                    "tipo": {
                        "$ref": "#/components/schemas/TramiteTipo"
                    },
                    "estado": {
                        "$ref": "#/components/schemas/TramiteEstado"
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "title": "Fecha Inicio"
                    },
                    "fecha_vencimiento": {
                        "type": "string",
                        "title": "Fecha Vencimiento"
                    },
                    "es_permanente": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Es Permanente"
                    },
                    "es_recurrente": {
                        "type": "boolean",
                        "title": "Es Recurrente"
                    },
                    "periodo_recurrencia": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/TramitePeriodoRecurrencia"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "observaciones": {
                        "items": {
                            "$ref": "#/components/schemas/Observacion"
                        },
                        "type": "array",
                        "title": "Observaciones",
                        "default": []
                    },
                    "documentos": {
                        "items": {
                            "$ref": "#/components/schemas/Documento"
                        },
                        "type": "array",
                        "title": "Documentos",
                        "default": []
                    },
                    "historial": {
                        "items": {
                            "$ref": "#/components/schemas/HistorialItem"
                        },
                        "type": "array",
                        "title": "Historial",
                        "default": []
                    },
                    "asignado_a": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Asignado A"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "tienda_id",
                    "nombre",
                    "tipo",
                    "estado",
                    "fecha_inicio",
                    "fecha_vencimiento",
                    "es_recurrente"
                ],
                "title": "Tramite"
            },
            "TramiteCreate": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "title": "Nombre"
                    },
                    "tipo": {
                        "type": "string",
                        "title": "Tipo"
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "title": "Fecha Inicio"
                    },
                    "fecha_vencimiento": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fecha Vencimiento"
                    },
                    "es_permanente": {
                        "type": "boolean",
                        "title": "Es Permanente",
                        "default": false
                    },
                    "es_recurrente": {
                        "type": "boolean",
                        "title": "Es Recurrente",
                        "default": false
                    },
                    "periodo_recurrencia": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Periodo Recurrencia"
                    }
                },
                "type": "object",
                "required": [
                    "nombre",
                    "tipo",
                    "fecha_inicio"
                ],
                "title": "TramiteCreate"
            },
            "TramiteEstado": {
                "type": "string",
                "enum": [
                    "pendiente_documentacion",
                    "en_revision",
                    "presentado",
                    "en_espera_resolucion",
                    "vigente",
                    "por_vencer",
                    "vencido"
                ],
                "title": "TramiteEstado"
            },
            "TramitePeriodoRecurrencia": {
                "type": "string",
                "enum": [
                    "anual",
                    "bianual"
                ],
                "title": "TramitePeriodoRecurrencia"
            },
...
            "TramiteTipo": {
                "type": "string",
                "enum": [
                    "federal",
                    "estatal",
                    "municipal"
                ],
                "title": "TramiteTipo"
            },
            "TramiteUpdate": {
                "properties": {
                    "nombre": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Nombre"
                    },
                    "fecha_inicio": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fecha Inicio"
                    },
                    "fecha_vencimiento": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fecha Vencimiento"
                    },
                    "es_permanente": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Es Permanente"
                    }
                },
                "type": "object",
                "title": "TramiteUpdate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/login"
                    }
                }
            }
        }
    }
}